api_django_drf/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
psycopg2==2.9.7
drf-yasg==1.21.5
PyJWT==2.8.0
